+++ lib/Backend/Definitions/Device/device_definition.dart
@@ -33,6 +33,43 @@ enum DeviceType {
   miniTail
 } //TODO extend with icon
 
+@HiveType(typeId: 14)
+enum EarSpeed {
+  @HiveField(1)
+  fast,
+  @HiveField(2)
+  slow,
+}
+
+extension EarSpeedExtension on EarSpeed {
+  String get name {
+    switch (this) {
+      case EarSpeed.fast:
+        return earSpeedFast();
+      case EarSpeed.slow:
+        return earSpeedSlow();
+    }
+  }
+
+  Widget get icon {
+    switch (this) {
+      case EarSpeed.fast:
+        return const Icon(Icons.fast_forward);
+      case EarSpeed.slow:
+        return const Icon(Icons.play_arrow);
+    }
+  }
+
+  String get command {
+    switch (this) {
+      case EarSpeed.fast:
+        return "SPEED FAST";
+      case EarSpeed.slow:
+        return "SPEED SLOW";
+    }
+  }
+}
+
 extension DeviceTypeExtension on DeviceType {
   String get name {
     switch (this) {