+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### V 1.0.9
 
+- Allow setting ear twist speed
 - Fix custom sounds on android and IOS.
 - Other Bug Fixes